+++ envvar.go
@@ -2,6 +2,7 @@ package env
 
 import (
 	"fmt"
+	"io"
 	"os"
 	"strings"
 )
@@ -16,7 +17,7 @@ const (
 	// ExitOnError will call os.Exit(2) if an error is found when parsing
 	ExitOnError
 
-	// PanicOnError will panic() if an error is found when parsing flags
+	// PanicOnError will panic() if an error is found when parsing environment variables
 	PanicOnError
 )
 
@@ -25,6 +26,7 @@ type EnvVarSet struct {
 	parsed        bool
 	vars          map[string]Value
 	errorHandling ErrorHandling
+	output        io.Writer // nil means stderr; use out() accessor
 }
 
 // NewEnvVarSet returns a new, empty environment variable set.
@@ -45,7 +47,30 @@ func (s *EnvVarSet) Var(value Value, name string, usage string) {
 		s.vars = make(map[string]Value)
 	}
 
+	_, alreadyThere := s.vars[name]
+	if alreadyThere {
+		msg := fmt.Sprintf("%s environment variable redefined: %s", name, name)
+
+		fmt.Fprintln(s.out(), msg)
+
+		panic(msg) // Happens only if environment variables are declared with identical names
+	}
+
 	s.vars[name] = value
+}
+
+func (s *EnvVarSet) out() io.Writer {
+	if s.output == nil {
+		return os.Stderr
+	}
+
+	return s.output
+}
+
+// SetOutput sets the destination for usage and error messages.
+// If output is nil, os.Stderr is used.
+func (s *EnvVarSet) SetOutput(output io.Writer) {
+	s.output = output
 }
 
 // Parse parses environment variables according to the definitions in the EnvVarSet.
@@ -62,7 +87,7 @@ func (s *EnvVarSet) Parse(environment map[string]string) error {
 				case ContinueOnError:
 					return err
 				case ExitOnError:
-					fmt.Println(err)
+					fmt.Fprintln(s.out(), err)
 					os.Exit(2)
 				case PanicOnError:
 					panic(err)