+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changes
+
+- `Var` panics when redeclaring an existing variable
+- Error output writer
+
 
 ## 0.1.0 - 2018-06-06